Make Helper#platform raise an exception on unknown OSs
Currently, #platform returns 'win' if it can't figure out what OS Ruby
is running on.

This leads to it attempting to run the Windows binary for
Chromedriver, and thus odd errors on FreeBSD.  See #1.

This change makes #platform raise an exception on unsupported OSs, to
make it clear what's going on.

FreeBSD support is a separate issue, which I will address if I have
the time.