Only ref Selenium if selenium-webdriver present
Before, the code was accessing the `Selenium::WebDriver::Chrome`
assuming that it had already been defined. There is no guarantee that
this will be the case.

Fix by explicitly loading the `selenium-webdriver` gem first in order to
define these constants. If the gem is not present, gracefully skip this
code instead of crashing with "uninitialized constant".

Fixes #60